Updated includes to use "/" instead of "\" Updated aspects of program to work with gcc

+++ src/MathHelper.h
@@ -14,7 +14,7 @@
 	#define _INFINITY 10000.0f
 #endif
 
-#define _PI_ 2.0f*std::acosf(0.0f)
+#define _PI_ 2.0f*std::acos(0.0f)
 
 //converts degrees to radians
 #define degToRad(d) d*(_PI_ / 180.0f)
